+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to Lumina Event Horizon are documented in this file.
 
-## [Unreleased]
+## [1.3.81] - 2026-08-17
 
 ### Fixed